Instead of MetricFu, run them individually
There is something weird in the dependencies of MetricFu. Additionally, a lot of the metrics are run individually.

The following metrics are run by separate rake tasks and don't need to be added:

* flay
* flog
* reek
* roodi
* rcov

I would not recommend using the following metrics in this project:

* Rails Best Practices is not appropriate
* churn is hopelessly outdated, the Gemfile uses exact versions for all dependencies...

Finally, re-adding saikuro will be handled in a separate commit.

JSON was only used in the Raketask for metric_fu and was therefore also removed from the Gemfile.