Update gemfile to use git source for adamantium
This change is not subjected to be merged to master. It makes it more
easy to cherry pick the problem and the solution indepentand from each
other.

Also this change removes git source for reek. No repository that has the
fix and that depends on a ruby2ruby version that is compatible with
heckle does exist anymore.